backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
pydantic-settings==2.2.1
httpx==0.27.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9 